Moved /react-components to match /twig folder structure
