+++ src/react-components/__tests__/setupEnzyme.ts
@@ -1,5 +1,5 @@
 
-import { configure } from 'enzyme';
+import { configure } from "enzyme";
 const Adapter = require("enzyme-adapter-react-16");
 
 configure({ adapter: new Adapter() });
