+++ src/react-components/components/02-molecules/menus/Breadcrumb/Breadcrumb.stories.tsx
@@ -1,0 +1,10 @@
+import * as React from "react";
+
+import Breadcrumb from "./Breadcrumb";
+import { action } from "@storybook/addon-actions";
+
+export default {
+  title: "Breadcrumb",
+};
+
+export const regular = () => <Breadcrumb items={[{ url: "#", text: "Home" }, { url: "#", text: "Parent" }]}></Breadcrumb>;
